Fallback to background color for missing or bad images
If the background image defined in the config file at `background:path`
cannot be found, render the lock screen background with the color
defined at `background:color`. If color is not defined fallback to
default color.

Added `valid` flag to `SPreloadTarget` and `SPreloadedAsset`.

If Cairo is unable to load an image, for whatever reason, the
PreloadTarget is invalidated. The Resource Gatherer will not attempt to
create a GL texture for invalid Preload assets and the resulting texture
will be invalid, allowing later render functions handle a bad texture as
they see fit. In this case a bad background texture is replaced with the
solid fill color.

This should allow safe fallback to a color without a bunch of Cairo and
GL buffers hanging around for no reason.